Refactor: Extract MovieDetailsScreen components
This commit refactors the `MovieDetailsScreen` by extracting its UI elements into smaller, reusable composables. This improves code organization, readability, and maintainability.

The large, monolithic screen composable has been broken down into the following specific components:
- `MovieHeroSection`: Displays the movie's backdrop, poster, title, and rating.
- `TrailersSection`: A self-contained card that shows a scrollable list of movie trailers.
- `SectionCard`: A generic card for displaying content sections with a title.
- `InfoItem`: A small composable for displaying an icon, a value, and a label (e.g., for runtime, language).
- `BoxOfficeItem`: A component for displaying financial data like budget and revenue.

All new components have been moved to the `features/movies/src/commonMain/kotlin/com/elna/moviedb/feature/movies/ui/movie_details/components/` package. The main `MovieDetailsScreen.kt` file now composes these smaller pieces, making the overall structure cleaner.
