+++ features/movies/src/commonMain/kotlin/com/elna/moviedb/feature/movies/ui/movie_details/components/BoxOfficeItem.kt
@@ -1,0 +1,37 @@
+package com.elna.moviedb.feature.movies.ui.movie_details.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+
+@Composable
+internal fun BoxOfficeItem(
+    label: String,
+    amount: Long
+) {
+    Row(
+        modifier = Modifier.fillMaxWidth(),
+        horizontalArrangement = Arrangement.SpaceBetween
+    ) {
+        Text(
+            text = label,
+            style = MaterialTheme.typography.bodyMedium,
+            fontWeight = FontWeight.Medium
+        )
+        Text(
+            text = "$${formatNumberWithCommas(amount)}",
+            style = MaterialTheme.typography.bodyMedium,
+            fontWeight = FontWeight.SemiBold,
+            color = MaterialTheme.colorScheme.primary
+        )
+    }
+}
+
+internal fun formatNumberWithCommas(number: Long): String {
+    return number.toString().reversed().chunked(3).joinToString(",").reversed()
+}
