+++ features/movies/src/commonMain/kotlin/com/elna/moviedb/feature/movies/ui/movie_details/components/MovieHeroSection.kt
@@ -1,0 +1,138 @@
+package com.elna.moviedb.feature.movies.ui.movie_details.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.systemBarsPadding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Star
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import com.elna.moviedb.core.model.MovieDetails
+import com.elna.moviedb.core.ui.utils.ImageLoader
+import com.elna.moviedb.resources.Res
+import com.elna.moviedb.resources.rating
+import com.elna.moviedb.resources.votes
+import org.jetbrains.compose.resources.stringResource
+
+@Composable
+internal fun MovieHeroSection(movie: MovieDetails) {
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(450.dp)
+    ) {
+        // Backdrop Image
+        movie.backdropPath?.let { backdropPath ->
+            ImageLoader(
+                imageUrl = backdropPath,
+                modifier = Modifier.fillMaxSize()
+            )
+        }
+
+        // Gradient Overlay
+        Box(
+            modifier = Modifier
+                .fillMaxSize()
+                .background(
+                    Brush.verticalGradient(
+                        colors = listOf(
+                            Color.Transparent,
+                            Color.Black.copy(alpha = 0.7f)
+                        )
+                    )
+                )
+        )
+
+        // Poster
+        movie.posterPath?.let { posterPath ->
+            Card(
+                modifier = Modifier
+                    .systemBarsPadding()
+                    .width(120.dp)
+                    .height(180.dp)
+                    .align(Alignment.TopStart)
+                    .padding(start = 16.dp),
+                shape = RoundedCornerShape(8.dp),
+                elevation = CardDefaults.cardElevation(defaultElevation = 8.dp)
+            ) {
+                ImageLoader(
+                    imageUrl = posterPath,
+                    modifier = Modifier.fillMaxSize()
+                )
+            }
+        }
+
+        // Title and Basic Info
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .align(Alignment.BottomStart)
+                .padding(16.dp),
+            verticalArrangement = Arrangement.spacedBy(8.dp)
+        ) {
+            Text(
+                text = movie.title,
+                style = MaterialTheme.typography.headlineLarge,
+                color = Color.White,
+                fontWeight = FontWeight.Bold
+            )
+
+            movie.tagline?.takeIf { it.isNotBlank() }?.let { tagline ->
+                Text(
+                    text = tagline,
+                    style = MaterialTheme.typography.bodyLarge,
+                    color = Color.White.copy(alpha = 0.8f),
+                    fontStyle = androidx.compose.ui.text.font.FontStyle.Italic
+                )
+            }
+
+            // Rating
+            movie.voteAverage?.let { rating ->
+                Row(
+                    verticalAlignment = Alignment.CenterVertically,
+                    horizontalArrangement = Arrangement.spacedBy(4.dp)
+                ) {
+                    Icon(
+                        imageVector = Icons.Default.Star,
+                        contentDescription = stringResource(Res.string.rating),
+                        tint = Color.Yellow,
+                        modifier = Modifier.size(24.dp)
+                    )
+                    Text(
+                        text = "${(rating * 10).toInt() / 10.0}",
+                        style = MaterialTheme.typography.bodyLarge,
+                        color = Color.White,
+                        fontWeight = FontWeight.SemiBold
+                    )
+                    movie.voteCount?.let { count ->
+                        Text(
+                            text = "($count ${stringResource(Res.string.votes)})",
+                            style = MaterialTheme.typography.bodyMedium,
+                            color = Color.White.copy(alpha = 0.7f)
+                        )
+                    }
+                }
+            }
+        }
+    }
+}
